pyproject.toml
[project]
name = "wiimote-control"
version = "0.1.0"
description = "Control your Mac with a Nintendo Wiimote — hands-free voice input via Wispr Flow"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyobjc-framework-applicationservices>=12.1",
    "pyobjc-framework-quartz>=12.1",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
]
